% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTPMFromBAMs.R
\name{calcTPMsFromBAMs}
\alias{calcTPMsFromBAMs}
\title{Calculate TPMs from BAM files}
\usage{
calcTPMsFromBAMs(bedOrdered, bamFiles, sampleNames, outputDir, mc.cores = 4)
}
\arguments{
\item{bedOrdered}{a ordered bed data table
tab separeated bed file containing capture design probes with the following 4 columns:
- chromosme
- start
- stop
- gene name *}

\item{bamFiles}{a vector of paths to bam files}

\item{sampleNames}{a vector of sample identifiers}

\item{outputDir}{a character value of output directory}

\item{mc.cores}{a numeric value of cores}
}
\description{
\code{calcTPMsFromBAMs()} is used to calculate TPM value from BAM file.
It takes in a vector of paths to bed files (bedFile), a vector of paths
to BAM files (bamFiles), a vector of sample identifiers (sampleNames),
an output directory (outputDir), and a numeric value of cores (mc.cores).
NOTE: Index files (.bai) should be available for each BAM file
}
