% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTxtoLine.R
\name{mapTxtoLine}
\alias{mapTxtoLine}
\title{mapTxtoLine(bed.gr, refGene.gr)}
\usage{
mapTxtoLine(bed.gr, refGene.gr)
}
\arguments{
\item{bed.gr}{a GRanges object that represents the genomic coordinates of a
set of probes.}

\item{refGene.gr}{a GRanges object that represents the genomic coordinates
of a set of reference transcripts, metatable includes a transcript_id column}
}
\value{
a list where each element corresponds to a transcript, and contains a
vector of line numbers in the probe file that correspond to that transcript.
}
\description{
The function first finds the overlaps between the probes and
the reference transcripts using the findOverlaps function.
It then extracts the transcript IDs from the reference transcripts that
overlap with the probes, and calculates the number of lines in the probe file
that correspond to each transcript.
}
