% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBed.R
\name{prepareBed}
\alias{prepareBed}
\title{prepareBed}
\usage{
prepareBed(bedFile, ref.genome)
}
\arguments{
\item{bedFile}{a path to capture design or exon file}

\item{ref.genome}{a path to indexed fa file or a BSgenome object}
}
\value{
The function returns the GRanges object with the metadata column
gcbias', excluding any intervals where the GC bias is not available.
}
\description{
takes in a bed file and a reference genome, and returns a
GRanges object with a metadata column 'gcbias'.
The function first checks if the reference genome is a valid path to a fasta
file or a BSgenome object. If it is not, the function throws an error.
Next, the function reads in the bed file using the fread function from the
data.table package. It then orders the bed file by chromosome and start
position, and removes any duplicated capture designs.
Finally, the function annotates the GC ratio for each interval in the bed
file. If the reference genome is a BSgenome object, the function uses the
getSeq function to extract the sequence for each interval, calculates the GC
content, and calculates the GC bias. If the reference genome is a fasta file,
the function uses the scanFa function from the Rsamtools package to extract
the sequence for each interval, calculates the GC content, and calculates
the GC bias.
}
