% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTPMData.R
\name{prepareTPMData}
\alias{prepareTPMData}
\title{prepareTPMData
read individual TPM bed file and prepare matrix for z-TPM calculation.}
\usage{
prepareTPMData(tpmFile, chrom = NULL, mc.cores = 2)
}
\arguments{
\item{tpmFile}{a vector of paths of tpm.bed file}

\item{chrom}{a vector of chromosome to read}

\item{mc.cores}{number of core for parallel running}
}
\value{
a data.table object with sample id named columns; each column stores
the TPM values
}
\description{
prepareTPMData
read individual TPM bed file and prepare matrix for z-TPM calculation.
}
