% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordertpmDt.R
\name{reordertpmDt}
\alias{reordertpmDt}
\title{Reorder BED and TPM matrix}
\usage{
reordertpmDt(bedFile, tpmDt)
}
\arguments{
\item{bedFile}{tab separeated bed file containing capture design probes with the following 4 columns:
- chromosme
- start
- stop
- gene name}

\item{tpmDt}{object returned by prepareTPMData}
}
\value{
a list of reordered bed and tmpDt
}
\description{
Reorder BED and TPM matrix
}
