% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegNormRDES.R
\name{SegNormRDES}
\alias{SegNormRDES}
\title{SegNormRDES: Performs segmentation on a given data frame using the specified segmentation method.}
\usage{
SegNormRDES(df, id, seg.method = "slm")
}
\arguments{
\item{df}{A data frame to be segmented.}

\item{id}{An identifier for the data.}

\item{seg.method}{A character string specifying the segmentation method to be used. Default is "slm" (Segmented Linear Model).}
}
\value{
A data table with the calculated values for each sequence.
}
\description{
SegNormRDES: Performs segmentation on a given data frame using the specified segmentation method.
}
\examples{
SegNormRDES(df, id, seg.method="slm")

}
