% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDupFinder-package.R
\docType{package}
\name{HMZDupFinder-package}
\alias{HMZDupFinder}
\alias{HMZDupFinder-package}
\title{HMZDupFinder: Detect intragenic homozygous duplication from large exome sequencing data.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Haowei Du \email{haoweid@bcm.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
